\name{MSeasyTkGUI-package}
\alias{MSeasyTkGUI-package}
\docType{package}
\title{
MSeasyTkGUI
}
\description{
MSeasyTkGUI is a user interface for MSeasy
manly based on code from ade4TkGUI
}
\details{
\tabular{ll}{
Package: \tab MSeasyTkGUI\cr
Type: \tab Package\cr
Version: \tab 5.3.1\cr
Date: \tab 2012-08-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Yann Guitton
Maintainer: Yann Guitton <yann.guitton@gmail.com>

}
\references{
ade4TkGUI was used as source code for MSeasyTkGUI
}

\keyword{ package }